Create new migrations for v1 and v2.
In order to upgrade timeclock on production from v1 to v2, the old migrations
directory was moved to migrations_v1 and new migrations were created for the v1
database and the v2 database.

Additionally, removed the automatic DB Models creation on create_app and updated
default database URL for DevelopmentConfig.